rearrange the instrument configuration file according to the new way
