+++ main.py
@@ -1,48 +1,10 @@
 from __future__ import division
 from instrumentino import InstrumentinoApp
-from instrumentino.components import Component
-from instrumentino.variables import AnalogVariablePercentage,\
-    DigitalVariableOnOff
+from instrumentino.components import AnalogInVariables, DigitalInOutVariables
+from instrumentino.variables import AnalogVariablePercentage
 from instrumentino.controllers.arduino import Arduino, ArduinoChannelIn_AnalolgInPin,\
-    ArduinoChannelInOut_DigitalPin, ArduinoChannelOut_DigitalPin
+    ArduinoChannelInOut_DigitalPin
 from instrumentino.screens.automation import Action
-from instrumentino.controllers import Controller
-import inspect
-import sys
-
-class AnalogInVariables(Component):
-    '''An array of analog input variables
-    '''
-    
-    def __init__(self, **kwargs):
-        ch_class = kwargs.get('ch_class', None)
-        controller = kwargs.get('controller', None)
-        channels_numbers = kwargs.get('channels_numbers', None)
-        sampling_rate = kwargs.get('sampling_rate', None)
-        
-        for i in channels_numbers:
-            ch_in = ch_class(controller=controller, number=i, sampling_rate=sampling_rate)
-            self.add_variable(AnalogVariablePercentage(name='Analog '+str(i), channel_in=ch_in))
-        
-        super(AnalogInVariables, self).__init__(**kwargs)
-
-
-class DigitalInOutVariables(Component):
-    '''An array of digital input/output variables
-    '''
-    
-    def __init__(self, **kwargs):
-        ch_class = kwargs.get('ch_class', None)
-        controller = kwargs.get('controller', None)
-        channels_numbers = kwargs.get('channels_numbers', None)
-        sampling_rate = kwargs.get('sampling_rate', None)
-        
-        for i in channels_numbers:
-            ch = ch_class(controller=controller, number=i, sampling_rate=sampling_rate)
-            self.add_variable(DigitalVariableOnOff(name='Digital '+str(i), channel_in=ch, channel_out=ch))
-
-        super(DigitalInOutVariables, self).__init__(**kwargs)
-
 
 # Define controllers
 arduino = Arduino()
